Fully separate nix from windows
